+++ CHANGELOG.md
@@ -31,6 +31,8 @@ The [history](HISTORY.md) contains the changelogs of older PostBOUND releases.
   table references on columns.
 - `util.simplify()` now supports single-item mappings as well.
 - `util.to_json()` now supports the standard date-like objects (_datetime_, _date_, _time_, and _timedelta_)
+- `util.write_df()` now automatically transforms complex objects in the data frame into their JSON representation before
+  writing.
 - In IPython and Jupyter sessions, common PostBOUND objects like SQL queries or query plans are now automatically
   pretty-printed. Instead of calling `print(plan.explain())`, one can now simply make `plan` the result of a cell.